+++ db/migrate/20250409064922_create_i_promocoes.rb
@@ -2,19 +2,19 @@
 
 class CreateIPromocoes < ActiveRecord::Migration[7.2]
   def up
-    create_table :i_promocoes do |t|
-      t.string   :nome
-      t.text     :descricao
-      t.boolean  :ativo
-      t.string   :slug
-      t.string   :created_by
-      t.string   :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:i_promocoes)
+      create_table :i_promocoes do |t|
+        t.string   :nome
+        t.text     :descricao
+        t.boolean  :ativo
+        t.string   :slug
+        t.datetime :deleted_at
+        t.timestamps
+      end
     end
   end
 
   def down
-    drop_table :i_promocoes
+    drop_table :i_promocoes if table_exists?(:i_promocoes)
   end
 end
