+++ db/migrate/20250421020452_create_i_itens_pedidos.rb
@@ -2,22 +2,25 @@
 
 class CreateIItensPedidos < ActiveRecord::Migration[7.2]
   def up
-    create_table :i_itens_pedidos do |t|
-      t.references :i_pedido,       null: false, foreign_key: true
-      t.references :i_produto,      null: false, foreign_key: true
-      t.integer    :quantidade
-      t.decimal    :preco_unitario, precision: 10, scale: 2
-      t.decimal    :subtotal,       precision: 10, scale: 2
-      t.datetime   :criado_em
-      t.datetime   :atualizado_em
-      t.string :created_by
-      t.string :updated_by
-      t.datetime :deleted_at
-      t.timestamps
+    unless table_exists?(:i_itens_pedidos)
+      create_table :i_itens_pedidos do |t|
+        t.references :i_pedido,       null: false, foreign_key: true
+        t.references :i_produto,      null: false, foreign_key: true
+        t.integer    :quantidade
+        t.decimal    :preco_unitario, precision: 10, scale: 2
+        t.decimal    :subtotal,       precision: 10, scale: 2
+        t.datetime   :criado_em
+        t.datetime   :atualizado_em
+        t.datetime   :deleted_at
+        t.timestamps
+      end
+
+      # 🔒 Índice único opcional: impede que o mesmo produto seja adicionado mais de uma vez no mesmo pedido
+      add_index :i_itens_pedidos, [:i_pedido_id, :i_produto_id], unique: true, name: 'index_itens_pedidos_unique'
     end
   end
 
   def down
-    drop_table :i_itens_pedidos
+    drop_table :i_itens_pedidos if table_exists?(:i_itens_pedidos)
   end
 end
